fixed feedback suggestions
